vc_utils.rb
